requirements.txt
discord.py>=2.4.0
pandas>=2.2.0
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.3
pyyaml>=6.0.0
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pytest>=8.2.0
